src/model/Priestess.java
package model;

/**
 * A playable hero class with the special skill: Heal.
 * @author Jonathan Abrams, Madison Pope, Martha Emerson
 */
public class Priestess extends Hero {
    /**
     * Construct a Priestess.
     * @param theName the provided name of the Hero.
     */
    public Priestess(final String theName) {
        super(theName);
    }

    /**
     * Perform special skill (HEAL)
     *
     * @return true if successful
     */
    @Override
    public boolean specialSkill(final DungeonCharacter theTarget) {
        heal(randomValue(myStats.minHeal(), myStats.maxHeal()));
        return true;
    }
}
